% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunction.R
\name{MVMRIE_EM}
\alias{MVMRIE_EM}
\title{Multivariable Mendelian Randomization with Incomplete Measurements on the Exposure Variables}
\usage{
MVMRIE_EM(IV.dat, pheno.dat, covar.dat = NULL, epsilon = 1e-05, num.ghq)
}
\arguments{
\item{IV.dat}{A required data frame containing data on the instrumental
variables (IVs) for each individual, with a header line indicating the names
of the IVs. Each row corresponds to one individual, and each column
corresponds to one IV. Missing values or \code{NA}s are not allowed.}

\item{pheno.dat}{A required data frame containing data on the outcome and the
two exposure variables, with a header line indicating their names. Missing
values or \code{NA}s are not allowed in this data frame. Each row corresponds to one individual,
and the individuals should be listed in the same order as in \code{IV.dat}.

The dataset \code{pheno.dat} should have five columns. The 1st column
contains the floating-point values for the continuous outcome. The 2nd column
(integer-valued) and the 3rd column (floating-point valued) contain
information on the first exposure variable, and the 4th column (integer-valued)
and the 5th column (floating-point valued) should be similarly
specified for the other exposure. See the "Details" section for instructions
on specifying these columns.}

\item{covar.dat}{An optional data frame containing data on the measured
covariates, with a header line indicating the names of these variables.
Each row corresponds to one individual, and the individuals should be listed
in the same order as in \code{IV.dat}. Each column corresponds to one
covariate. Missing values or \code{NA}s are not allowed. The default value is \code{NULL}.}

\item{epsilon}{A positive floating-point value indicating the convergence
threshold of the expectation-maximization (EM) algorithm. The iterations will be terminated if the
Euclidean distance between the parameter values at two successive iterations
is less than this value. The default value is \code{1e-5}.}

\item{num.ghq}{A positive integer indicating the number of quadrature nodes
to be used when approximating the integrals with Gauss-Hermite quadrature
during the computation.}
}
\value{
A list with the following components will be returned:
\itemize{
\item{\code{results_reg} contains the point estimates for
\mjeqn{\alpha_1, \alpha_2, \beta, \gamma_1}{ASCII representation}, and
\mjeqn{\gamma_2}{ASCII representation}. Standard error estimates and p-values
are also provided. The inference on the direct causal effects of interest is
presented in the last two rows.}
\item{\code{results_var} contains the estimates for the variance components
(i.e., the variances of \mjeqn{\epsilon_1, \epsilon_2}{ASCII representation},
and \mjeqn{\epsilon_Y}{ASCII representation} and their pairwise correlation).}
\item{\code{full_cov_mat} is the entire estimated covariance matrix
derived from the Louis formula (Little & Rubin, 2019).}
}
}
\description{
This function is used for multivariable Mendelian randomization (MVMR) with
two continuous exposure variables and a continuous outcome, where the exposure
variables are potentially unmeasured and subject to detection limits.
}
\details{
\loadmathjax
\strong{1) Model and Method}

Let \mjeqn{Y}{ASCII representation} be a continuous outcome,
\mjeqn{S_1}{ASCII representation} and \mjeqn{S_2}{ASCII representation} be
two continuous exposure variables that are potentially unmeasured and subject
to detection limits, \mjeqn{\mathbf{G}}{ASCII representation} be a vector of
IVs for \mjeqn{S_1}{ASCII representation} and \mjeqn{S_2}{ASCII representation},
and \mjeqn{\mathbf{Z}}{ASCII representation} be a vector of measured
covariates. Assume that the first component of \mjeqn{\mathbf{Z}}{ASCII representation}
is 1, and let \mjeqn{\mathbf{X} = (\mathbf{G}^T, \mathbf{Z}^T)^T}{ASCII representation}.
Consider the following models:
\mjdeqn{S_1 = \alpha_1^T \mathbf{X} + \epsilon_1,}{ASCII representation}
\mjdeqn{S_2 = \alpha_2^T \mathbf{X} + \epsilon_2,}{ASCII representation}
\mjdeqn{Y = \gamma_1 S_1 + \gamma_2 S_2 + \beta^T \mathbf{Z} + \epsilon_Y,}{ASCII representation}
where \mjeqn{\alpha_1}{ASCII representation}, \mjeqn{\alpha_2}{ASCII representation},
and \mjeqn{\beta}{ASCII representation} are regression parameters,
\mjeqn{\gamma_1}{ASCII representation} and \mjeqn{\gamma_2}{ASCII representation}
represent the direct causal effect of \mjeqn{S_1}{ASCII representation} and
\mjeqn{S_2}{ASCII representation} on \mjeqn{Y}{ASCII representation},
respectively, and \mjeqn{(\epsilon_1, \epsilon_2, \epsilon_Y)^T}{ASCII representation}
is a three-dimensional normal random vector with mean zero and an
unstructured covariance matrix. We treat \mjeqn{S_1}{ASCII representation}
and \mjeqn{S_2}{ASCII representation} as potentially missing data. The
estimations are carried out using the maximum likelihood estimation method,
and the EM algorithm is used to handle the incomplete measurements on the exposures.

\strong{2) The \code{pheno.dat} argument}

For the argument \code{pheno.dat}, the \mjeqn{i}{ASCII representation}th
element of the 2nd column indicates whether the first exposure
\mjeqn{S_1}{ASCII representation} of the \mjeqn{i}{ASCII representation}th
subject is observed, beyond detection limits, or unmeasured, and the
corresponding value in the 3rd column should be the exact measurement of
\mjeqn{S_1}{ASCII representation}, the detection limit, and a dummy value of
-999, respectively. Below is a table showing how the 2nd and the 3rd column
should be specified when \mjeqn{S_1}{ASCII representation} is observed,
beyond detection limits, or unmeasured.\tabular{lcc}{
    \tab The 2nd column \tab The 3rd column \cr
   Measured and detectable \tab 0 \tab the exact measurement \cr
   Measured but below the lower detection limit \tab 1 \tab the lower detection limit \cr
   Measured but above the upper detection limit \tab 2 \tab the upper detection limit \cr
   Unmeasured \tab 3 \tab -999 \cr
}


The 4th and 5th columns of \code{pheno.dat} should be specified in
the same way for \mjeqn{S_2}{ASCII representation}.
}
\references{
Little, R. J., & Rubin, D. B. (2019). Statistical analysis with missing data (3rd Edition). John Wiley & Sons.
}
